aya-ebpf: Use `i32` as a type for eBPF helper return codes
The helpers always return a signed 64-bit r0 value, but the JIT that
translates eBPF into native instructions differs by architecture. On
x86_64 the generated code writes the helper result into a 64-bit
register, so the CPU sign-extends negative errnos automatically. On
aarch64 the JIT frequently uses 32-bit operations (w0) when copying the
helper return and only zero-extends into the upper half of x0.

That results in broken errno codes on aarch64. For example, when a map
operation returns `-ENOENT`, which is supposed to be -2, the i64
representation yields something like `0x0000_0000_FFFF_FFFE`
(4294967294) instead of -2. In short: the ARM64 JIT doesn’t preserve the
sign in the upper half of the 64-bit register, and the error code has to
be cast to a 32-bit integer to make the sign visible.

This makes it awkward for users, because that behavior prevents simply
comparing helper error codes with constants like `libc::ENOENT` without
manual casts.

Given that the maximum error code limit `MAX_ERRNO` in the kernel is
4095, which fits in a 32-bit integer, coerce the error codes to be `i32`
and make all helpers return `Result<T, i32>`.

+++ ebpf/aya-ebpf/src/btf_maps/sk_storage.rs
@@ -1,7 +1,6 @@
 use core::ptr;
 
 use aya_ebpf_bindings::bindings::{BPF_F_NO_PREALLOC, BPF_SK_STORAGE_GET_F_CREATE, bpf_sock};
-use aya_ebpf_cty::c_long;
 
 use crate::{
     btf_maps::btf_map_def,
@@ -68,8 +67,8 @@ impl<T> SkStorage<T> {
     ///
     /// This function may dereference the pointer `sk`.
     #[inline(always)]
-    pub unsafe fn delete(&self, sk: *mut bpf_sock) -> Result<(), c_long> {
+    pub unsafe fn delete(&self, sk: *mut bpf_sock) -> Result<(), i32> {
         let ret = unsafe { bpf_sk_storage_delete(self.as_ptr(), sk.cast()) };
-        if ret == 0 { Ok(()) } else { Err(ret) }
+        if ret == 0 { Ok(()) } else { Err(ret as i32) }
     }
 }
